Refactor encoding code into separate package
