+++ testdata/yamltothrift/binary_data.json
@@ -1,7 +1,7 @@
 {
   "arg": {
     "b": {
-      "file": "testdata/yamltothrift/binary_data.file"
+      "file": "../testdata/yamltothrift/binary_data.file"
     }
   }
 }
